March 20, 2017
Implemented an initial navigation strategy for Kmeans

Edited files:
- bepoppy8.xml airframe
- bepoppy8.xml flightplan
- bepoppy8.c/bepoppy8.h
- bepoppy8_cv.cpp/bepoppy8_cv.h

No new files were created.

The code that comes with this commit is of preliminary nature (lots of if statements, etc.)
Moreover, an additional navigation functions needs to be created that allows the Bebop to
adjust his heading at an obstruction encounter, without coming to a full stop (as is done
with set_waypoint_here).

Author: Tijmen Pollack